Summary: Connect management between the beanstalkd proxy and the server
Fixes: https://github.com/1xyz/coolbeans/issues/24

Problem: Problem: We have a three node cluster, and when have a beanstalkd server which acts a client to this service. When the leader shuts down, we need the beanstalkd server (the proxy client) to automatically switch over to using the correct leader. In this way a client that is connected to beanstalkd proxy-server does not necessarily see this disruption.

High level approach:

We use the health check feature of the GRPC load-balancer, only the leader reports healthy. When the leader switches the proxy client reconnects to the new leader.